LayoutTests/editing/async-clipboard/clipboard-wrapper-stays-alive-expected.txt

Verifies that the wrapper for Navigator's Clipboard remains consistent across GC passes. This test must be run using WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS subframeNavigator.clipboard.foo is "bar"
PASS successfullyParsed is true

TEST COMPLETE
